fix var format
